Rebuild Zulip import/export tool to actually work.
The previous export tool would only work properly for small realms,
and was missing a number of important features:
* Export of avatars and uploads
* Export of presence data, activity data, etc.
* Faithful export/import of timestamps
* Parallel export of messages
* Not OOM killing for large realms

The new tool runs as a pair of documented management commands, and
solves all of those problems.

Also we add a new management command for exporting the data of an
individual user.